+++ test/geocoder/googlegeocoder.js
@@ -38,23 +38,23 @@
         });
 
         describe('#geocode' , function() {
-            it('Should not accept Ipv4', function() {
+            it('Should not accept IPv4', function() {
 
                 var googleAdapter = new GoogleGeocoder(mockedHttpAdapter);
 
                 expect(function() {
                         googleAdapter.geocode('127.0.0.1');
-                }).to.throw(Error, 'GoogleGeocoder no suport geocoding IPv4');
+                }).to.throw(Error, 'GoogleGeocoder does not support geocoding IPv4');
 
             });
 
-            it('Should not accept Ipv6', function() {
+            it('Should not accept IPv6', function() {
 
                 var googleAdapter = new GoogleGeocoder(mockedHttpAdapter);
 
                 expect(function() {
                         googleAdapter.geocode('2001:0db8:0000:85a3:0000:0000:ac1f:8001');
-                }).to.throw(Error, 'GoogleGeocoder no suport geocoding IPv6');
+                }).to.throw(Error, 'GoogleGeocoder does not support geocoding IPv6');
 
             });
 
@@ -121,7 +121,7 @@
                 mock.verify();
             });
 
-            it('Should return geocoded adress', function(done) {
+            it('Should return geocoded address', function(done) {
                 var mock = sinon.mock(mockedHttpAdapter);
                 mock.expects('get').once().callsArgWith(2, false, { status: "OK", results: [{
                         geometry: {location : {
@@ -204,7 +204,7 @@
 
             });
 
-            it('Should return geocoded adress', function(done) {
+            it('Should return geocoded address', function(done) {
                 var mock = sinon.mock(mockedHttpAdapter);
                 mock.expects('get').once().callsArgWith(2, false, { status: "OK", results: [{
                         geometry: {location : {